% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_f.R
\name{partial_f}
\alias{partial_f}
\title{partial_f calculates fishing mortality by fleet}
\usage{
partial_f(fit, partial_stocks, ...)
}
\arguments{
\item{fit}{An object of class a4aFitSA with the assessment fit (no default)}

\item{partial_stocks}{An object of class FLStocks containing the partial stock objects (fleets)}
}
\value{
FLQuants
}
\description{
Produces a FLQuant with fishing mortalities by unit.
}
