% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderHtmla4a.R
\name{renderReporta4a}
\alias{renderReporta4a}
\title{Generate a standard HTML report for a4a results}
\usage{
renderReporta4a(
  stk,
  idx,
  fit,
  comp_stk = NA,
  output_dir = getwd(),
  file_name = "a4a_report.html"
)
}
\arguments{
\item{stk}{An FLStock object}

\item{idx}{An FLIndices object}

\item{fit}{An a4aFit object}

\item{comp_stk}{Optional FLStock object to compare against. Defaults to NA.}

\item{output_dir}{Directory to save the HTML report (defaults to current working directory)}

\item{file_name}{Name of the output HTML file}
}
\value{
The file path to the generated HTML report
}
\description{
Generate a standard HTML report for a4a results
}
