% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_next_year.R
\name{ssb_next_year}
\alias{ssb_next_year}
\title{ssb_next_year calculates the survivors at the beginning of the year after your last year in the assessment}
\usage{
ssb_next_year(stock, rec.years = 10, biol.years = 3)
}
\arguments{
\item{stock}{An object of class FLStock with the assessment results (no default)}

\item{rec.years}{Number of years to use for the geometric mean recruitment assumption (default: 10)}

\item{biol.years}{Number of years to use for the mean of the biological parameters (m, stock.wt, mat) (default: 3)}
}
\value{
FLStock.
}
\description{
Produces an FLStock that can be used to plot the SSB in the year after your final year. To be used only if harvest.spwn and m.spwn are 0. Note that in the stock summary plot the recruitment assumption will also be shown.
}
\examples{
data("ple4")
ple4_new <-ssb_next_year(ple4)
plot(ple4_new)

ple4_new1 <-ssb_next_year(ple4,3,1)
plot(ple4_new1)
}
